% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coauthors.R
\name{get_coauthors}
\alias{get_coauthors}
\title{Get list of co-authors}
\usage{
get_coauthors(name, orcid, source = c("crossref"))
}
\arguments{
\item{name}{(required) Name of person for which you want the co-authors. Can
be a \link{person} object or a character string that will be coerced as a \link{person}
object.}

\item{orcid}{(optional) ORCiD of the person for which you want to co-authors.
If provided, only articles where the author (\code{name}) added their ORCiD will
be used to return the list of co-authors. This might result in the omission
of some co-authors.}

\item{source}{(required) The data source. Currently only \code{"crossref"}}
}
\description{
Get list of co-authors
}
