% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coauthors_crossref.R
\name{get_coauthors_crossref}
\alias{get_coauthors_crossref}
\title{Get list of co-authors from CrossRef}
\usage{
get_coauthors_crossref(name, orcid = NULL)
}
\arguments{
\item{name}{(required) Name of person for which you want the co-authors. Can
be a \link{person} object or a character string that will be coerced as a \link{person}
object.}

\item{orcid}{(optional) ORCiD of the person for which you want to co-authors.
If provided, only articles where the author (\code{name}) added their ORCiD will
be used to return the list of co-authors. This might result in the omission
of some co-authors.}
}
\description{
Get list of co-authors from CrossRef
}
